requirements/base.txt
# Core Django
Django>=5.0,<5.1
psycopg2-binary>=2.9.9
python-decouple>=3.8
Pillow>=10.2.0

# Security - Bcrypt for password hashing
bcrypt>=4.1.2

# Django REST Framework
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.1
django-filter>=24.0
drf-spectacular>=0.27.0

# CORS
django-cors-headers>=4.3.1

# Celery & Redis
celery>=5.3.6
redis>=5.0.1
django-celery-beat>=2.5.0

# Cloud Storage
cloudinary>=1.38.0
django-cloudinary-storage>=0.3.0

# Utilities
python-dateutil>=2.8.2
pytz>=2024.1
